Make FdbCApi functions private and add [[nodiscard]] wrappers
We recently had a bug where the result of an FdbCApi function was
incorrectly discarded. We can't make function pointers [[nodiscard]], so
instead we'll wrap them in functions and make them private.

This change also fixes the ensuing [[nodiscard]] warnings, and moves
existing checks for whether or not the function pointer is set inside
the new wrapper functions.